Fjern toggle for registrering Hentes fra veilarbperson.
