+++ src/mock/api/veilarbpersonflatefs.ts
@@ -5,16 +5,14 @@ import {
 	PERSONALIA_DATA_FRA_PDL,
 	PERSONALIA_DATA_FRA_TPS,
 	SPOR_OM_TILBAKEMELDING,
-	INNSATSGRUPPE_OG_HOVEDMAL_FRA_VEDTAKSSTOTTE,
-	HENT_REGISTRERING_FRA_VEILARBPERSON,
+	INNSATSGRUPPE_OG_HOVEDMAL_FRA_VEDTAKSSTOTTE
 } from '../../rest/datatyper/feature';
 
 const features: Features = {
 	[PERSONALIA_DATA_FRA_PDL]: true,
 	[PERSONALIA_DATA_FRA_TPS]: false,
 	[SPOR_OM_TILBAKEMELDING]: true,
-	[INNSATSGRUPPE_OG_HOVEDMAL_FRA_VEDTAKSSTOTTE]: false,
-	[HENT_REGISTRERING_FRA_VEILARBPERSON]: true
+	[INNSATSGRUPPE_OG_HOVEDMAL_FRA_VEDTAKSSTOTTE]: false
 };
 
 export const veilarbpersonflatefsHandlers: RequestHandlersList = [
