+++ src/rest/api.ts
@@ -19,9 +19,8 @@ const headers = {
 	headers: { 'Nav-Consumer-Id': APP_NAME }
 };
 
-export const useFetchRegistrering = (fnr: string, hentfraVeilarbPerson: boolean = false) => {
-	const url = hentfraVeilarbPerson ? `/veilarbperson/api/person/registrering?fnr=${fnr}` : `/veilarbregistrering/api/registrering?fnr=${fnr}`;
-	return useFetch<RegistreringsData>(url, headers);
+export const useFetchRegistrering = (fnr: string) => {
+	return useFetch<RegistreringsData>(`/veilarbperson/api/person/registrering?fnr=${fnr}`, headers);
 };
 
 export const useFetchCvOgJobbprofil = (fnr: string) =>
